feat: Add Mailchimp SDK dependency and mailchimp configuration
This commit adds the "mailchimp/marketing" package as a dependency in composer.json. It also includes the necessary changes in composer.lock to reflect the addition of the package. Additionally, it adds a new configuration entry for Mailchimp in services.php. Finally, it modifies layout.blade.php to include smooth scrolling behavior and an anchor link for newsletter subscription.